Fix NameError in Python 3
This commit fixes the NameError produced by calling
 filter(lambda string: isinstance(string, basestring))
in Python 3. Because filter() is a generator, the lambda function doesn't
actually called at the point of the filter() statement (which is within
a try/except block), it gets called later and raises the error outside
the try block. With this commit, the code now uses six to identify the
relevant string type, preventing any errors from being raised in an
uncontrolled context.